requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
resend==2.5.0
pdfplumber==0.11.4
rapidocr-onnxruntime==1.4.4
Pillow==11.1.0
